requirements.txt
# BigQuery AI Application Dependencies

# Core BigQuery and Google Cloud
google-cloud-bigquery>=3.11.0
google-cloud-storage>=2.10.0
google-cloud-aiplatform>=1.35.0
google-cloud-functions>=1.8.0
google-cloud-logging>=3.8.0
google-cloud-monitoring>=2.15.0

# BigFrames for BigQuery AI
bigframes>=1.0.0

# Web Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
pydantic>=2.5.0

# AI and ML
scikit-learn>=1.3.0
tensorflow>=2.13.0
torch>=2.1.0

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0
aiofiles>=23.2.0
python-multipart>=0.0.6

# Logging and Monitoring
structlog>=23.2.0
prometheus-client>=0.19.0

# Security
cryptography>=41.0.0
passlib>=1.7.4
python-jose[cryptography]>=3.3.0

# Development and Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
httpx>=0.25.0

# Code Quality
black>=23.10.0
flake8>=6.1.0
isort>=5.12.0
mypy>=1.7.0
